+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [4.15.1](https://github.com/crowdin/crowdin-cli/compare/4.15.0...4.15.1) (2026-08-20)
+
+
+### Bug Fixes
+
+* **context:** detect the AI context section by its markers ([#1074](https://github.com/crowdin/crowdin-cli/issues/1074)) ([5a2a99e](https://github.com/crowdin/crowdin-cli/commit/5a2a99e98f08d2b0f6e5ef8b80c12d05c00e7456))
+* trim dest ** prefix only at a path-segment boundary ([#1075](https://github.com/crowdin/crowdin-cli/issues/1075)) ([a2c0b43](https://github.com/crowdin/crowdin-cli/commit/a2c0b436924d3970b084e62898a793d923cdd590))
+
+# [4.15.0](https://github.com/crowdin/crowdin-cli/compare/4.14.4...4.15.0) (2026-07-23)
+
+
+### Features
+
+* file context update ([#1053](https://github.com/crowdin/crowdin-cli/issues/1053)) ([8d70b70](https://github.com/crowdin/crowdin-cli/commit/8d70b70488679d6ab3afc76b33586816ad8e68de))
+* show error message on build fail ([#1051](https://github.com/crowdin/crowdin-cli/issues/1051)) ([ea76cf0](https://github.com/crowdin/crowdin-cli/commit/ea76cf026bdda6b85a9d76485dc2b44d926ae58a))
+
 ## [4.14.4](https://github.com/crowdin/crowdin-cli/compare/4.14.3...4.14.4) (2026-07-15)